优化Appdelegate
