+++ ShadowsocksX-NG/Base.lproj/MainMenu.xib
@@ -12,7 +12,7 @@
         </customObject>
         <customObject id="-1" userLabel="First Responder" customClass="FirstResponder"/>
         <customObject id="-3" userLabel="Application" customClass="NSObject"/>
-        <customObject id="Voe-Tx-rLC" customClass="AppDelegate" customModule="ShadowsocksX_NG" customModuleProvider="target">
+        <customObject id="Voe-Tx-rLC" customClass="MainMenuManager" customModule="ShadowsocksX_NG" customModuleProvider="target">
             <connections>
                 <outlet property="ACLAutoModeMenuItem" destination="sDG-YV-yuQ" id="kIE-Ua-zvx"/>
                 <outlet property="ACLBackChinaMenuItem" destination="lam-el-r4I" id="O41-Lr-e0h"/>
@@ -42,7 +42,7 @@
                 <outlet property="serversPreferencesMenuItem" destination="M5r-E7-44f" id="voe-SX-k6a"/>
                 <outlet property="showBunchJsonExampleFileItem" destination="pdy-JE-50Q" id="aWi-DB-deq"/>
                 <outlet property="showQRCodeMenuItem" destination="R6A-96-Zcb" id="XHz-pz-nCz"/>
-                <outlet property="statusMenu" destination="Hob-KD-bx9" id="clA-ZW-0pT"/>
+                <outlet property="statusMenu" destination="Hob-KD-bx9" id="Z0D-Ny-SwF"/>
                 <outlet property="tcpMenuItem" destination="eL4-tv-vDa" id="AMk-hV-HJ3"/>
                 <outlet property="toggleRunningMenuItem" destination="GSu-Tf-StS" id="XHw-pU-nCa"/>
                 <outlet property="updateSubscribeAtLaunchMenuItem" destination="nQ4-Nv-tox" id="dOD-xa-Rw4"/>
@@ -308,25 +308,11 @@
                 <menuItem isSeparatorItem="YES" id="lSf-ni-kjo"/>
                 <menuItem title="Quit" keyEquivalent="q" id="vJS-JW-byz">
                     <connections>
-                        <action selector="terminate:" target="-2" id="Mrp-4z-omR"/>
+                        <action selector="quitApp:" target="Voe-Tx-rLC" id="Rh0-dZ-wfQ"/>
                     </connections>
                 </menuItem>
             </items>
             <point key="canvasLocation" x="140" y="152"/>
-        </menu>
-        <menu id="2oY-e9-q1Z">
-            <items>
-                <menuItem title="Item 1" id="oR3-cT-RTC">
-                    <modifierMask key="keyEquivalentModifierMask"/>
-                </menuItem>
-                <menuItem title="Item 2" id="lay-XB-0oV">
-                    <modifierMask key="keyEquivalentModifierMask"/>
-                </menuItem>
-                <menuItem title="Item 3" id="m00-ru-6Po">
-                    <modifierMask key="keyEquivalentModifierMask"/>
-                </menuItem>
-            </items>
-            <point key="canvasLocation" x="247.5" y="482"/>
         </menu>
     </objects>
 </document>
