+++ ShadowsocksX-NG/Notifications.swift
@@ -14,3 +14,5 @@ let NOTIFY_ADV_PROXY_CONF_CHANGED = Notification.Name(rawValue: "NOTIFY_ADV_PROX
 let NOTIFY_HTTP_CONF_CHANGED = Notification.Name(rawValue: "NOTIFY_HTTP_CONF_CHANGED")
 let NOTIFY_INVALIDE_QR = Notification.Name(rawValue: "NOTIFY_INVALIDE_QR")
 let NOTIFY_FOUND_SS_URL = Notification.Name(rawValue: "NOTIFY_FOUND_SS_URL")
+let NOTIFY_UPDATE_MAINMENU = Notification.Name(rawValue: "NOTIFY_UPDATE_MAINMENU")
+let NOTIFY_TOGGLE_RUNNING = Notification.Name(rawValue: "NOTIFY_TOGGLE_RUNNING")
