+++ ShadowsocksX-NG/PingClient.swift
@@ -120,8 +120,7 @@ class PingServers:NSObject{
                 
                 DispatchQueue.main.async {
                     isTesting = false
-                    (NSApplication.shared.delegate as! AppDelegate).updateServersMenu()
-                    (NSApplication.shared.delegate as! AppDelegate).updateRunningModeMenu()
+                    NotificationCenter.default.post(name: NOTIFY_UPDATE_MAINMENU, object: nil)
                 }
             }
         }
